break out tests for dev extension to main it easier to maintain
